Cleanup Option,OptUtil,dialog interactions/code.
When dialog commits change, it now directly does
opt.setValue; if oldValue==oldValue setValue now
does nothing, in particular fire events.
It used to setPreference and wait
for the pref event to change the value. This could
mainifest as out of date info in dialog.

Add synronization protection to Option [sg]etValue;
note that preferences are often async.

Code simplification.